Switch between scalar and vector displays
Check the first feature in the data to determine whether we're dealing
with vector or scalar data (this makes an assumption that all features
in the collection have the same type) and then select either a vector or
scalar display.

The scalar display is just the existing display component, although it
needs some fixes for showing scalar data on the map. The vector display
is just a test component to ensure that the dynamic selection of the
component is working.
