+++ services/ui/src/js/components/DiagnosticView/DiagnosticView.js
@@ -1,24 +1,19 @@
 import { html } from "htm/preact";
 import { useCallback, useEffect, useState } from "preact/hooks";
 
-import useBrushedScalar from "./useBrushedScalar";
-import VariableDisplay from "./VariableDisplay";
+import ScalarVariable from "./ScalarVariable";
+import VectorVariable from "./VectorVariable";
 
 export default function DiagnosticView() {
   const [display, setDisplay] = useState(null);
   const [variables, setVariables] = useState({});
-  const [selection, setSelection] = useState(null);
   const [featureCollection, setFeatureCollection] = useState({ features: [] });
-  const [guess, setGuessBrush] = useBrushedScalar({
-    features: featureCollection.features,
-    loop: "guess",
-    variable: "magnitude",
-  });
-  const [analysis, setAnalysisBrush] = useBrushedScalar({
-    features: featureCollection.features,
-    loop: "analysis",
-    variable: "magnitude",
-  });
+
+  useEffect(() => {
+    fetch("/api/diag/")
+      .then((response) => response.json())
+      .then((json) => setVariables(json));
+  }, []);
 
   useEffect(() => {
     if (!display) return;
@@ -27,23 +22,14 @@ export default function DiagnosticView() {
       .then((json) => setFeatureCollection(json));
   }, [display]);
 
-  useEffect(
-    () =>
-      fetch("/api/diag/")
-        .then((response) => response.json())
-        .then((json) => setVariables(json)),
-    []
-  );
-
-  const brushCallback = (event) => {
-    setSelection(event.detail);
-    setGuessBrush(event.detail);
-    setAnalysisBrush(event.detail);
-  };
-
   const onVariableSelect = useCallback((event) => {
     setDisplay(event.target.value);
-  });
+  }, []);
+
+  const VariableComponent =
+    featureCollection.features[0]?.properties?.type === "vector"
+      ? VectorVariable
+      : ScalarVariable;
 
   const options = Object.entries(variables).map(
     ([name, url]) => html`<option value=${url}>${name}</option>`
@@ -54,22 +40,6 @@ export default function DiagnosticView() {
       ${options}
     </select>
 
-    <h2>Guess</h2>
-    <${VariableDisplay}
-      loop="guess"
-      distribution=${guess}
-      observations=${featureCollection}
-      selection=${selection}
-      brushCallback=${brushCallback}
-    />
-
-    <h2>Analysis</h2>
-    <${VariableDisplay}
-      loop="analysis"
-      distribution=${analysis}
-      observations=${featureCollection}
-      selection=${selection}
-      brushCallback=${brushCallback}
-    />
+    <${VariableComponent} featureCollection=${featureCollection} />
   </div>`;
 }
