+++ services/ui/src/js/components/DiagnosticView/VectorVariable.js
@@ -1,0 +1,5 @@
+import { html } from "htm/preact";
+
+export default function VectorVariable() {
+  return html`<p>Test</p>`;
+}
